Remove unit tests because node cannot be mocked for these tests.
