+++ spec/system/rbeapi/client_spec.rb
@@ -1,3 +1,34 @@
+#
+# Copyright (c) 2015, Arista Networks, Inc.
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met:
+#
+#   Redistributions of source code must retain the above copyright notice,
+#   this list of conditions and the following disclaimer.
+#
+#   Redistributions in binary form must reproduce the above copyright
+#   notice, this list of conditions and the following disclaimer in the
+#   documentation and/or other materials provided with the distribution.
+#
+#   Neither the name of Arista Networks nor the names of its
+#   contributors may be used to endorse or promote products derived from
+#   this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL ARISTA NETWORKS
+# BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
+# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
+# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR
+# BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+# WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE
+# OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN
+# IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+#
 require 'spec_helper'
 
 require 'rbeapi/client'
@@ -64,35 +95,39 @@ describe Rbeapi::Client do
   end
 
   # Client class methods
-  describe '#connect_to' do
-    it 'retrieves the node config' do
-      expect(node).to be_kind_of(Rbeapi::Client::Node)
-    end
-  end
-
-  describe '#config' do
-    it 'returns the currently loaded config object' do
-      expect(subject.config.read(dut_conf)).to eq(transport: 'socket')
-      expect(subject.connect_to('dut')).to be_kind_of(Rbeapi::Client::Node)
-    end
-  end
-
   describe '#config_for' do
-    it 'returns the configuration options for veos01' do
-      expect(subject.config.read(test_conf)).to eq(nil)
+    it 'returns the configuration options for the connection' do
+      expect(subject.load_config(test_conf)).to eq(nil)
       expect(subject.config_for('veos01')).to eq(veos01)
     end
 
-    it 'returns the configuration options for veos05' do
-      expect(subject.config.read(test_conf)).to eq(nil)
-      expect(subject.config_for('veos05')).to eq(veos05)
+    it 'returns nil if connection does not exist' do
+      expect(subject.config_for('veos22')).to eq(nil)
+    end
+  end
+
+  describe '#connect_to' do
+    it 'retrieves the node config' do
+      expect(subject.connect_to('veos01')).to be_truthy
+    end
+
+    it 'returns nil if connection does not exist' do
+      expect(subject.connect_to('veos22')).to eq(nil)
     end
   end
 
   describe '#load_config' do
     it 'overrides the default conf file loaded in the config' do
       expect(subject.load_config(test_conf)).to eq(nil)
+    end
+
+    it 'returns nil if connection does not exit' do
+      expect(subject.load_config(test_conf)).to eq(nil)
       expect(subject.config_for('dut')).to eq(nil)
+    end
+
+    it 'returns conf settings if connection exists' do
+      expect(subject.load_config(test_conf)).to eq(nil)
       expect(subject.config_for('veos01')).to eq(veos01)
     end
   end
@@ -100,13 +135,19 @@ describe Rbeapi::Client do
   # Config class methods
   describe 'config' do
     it 'gets the loaded configuration file data' do
+      expect(subject.load_config(test_conf)).to eq(nil)
       expect(subject.config.to_s).to include(test_data[0])
     end
   end
 
   describe '#read' do
-    it 'read the specified filename and load it' do
-      expect(subject.load_config(dut_conf)).to eq(transport: 'socket')
+    it 'read the specified filename and load dut' do
+      expect(subject.config.read(dut_conf)).to eq(transport: 'socket')
+      expect(subject.config.to_s)
+        .to include('host', 'username', 'password', '[connection:dut]')
+    end
+
+    it 'read the specified filename and load test' do
       expect(subject.config.read(test_conf)).to eq(nil)
       expect(subject.config.to_s).to include(test_data[0])
     end
@@ -124,11 +165,11 @@ describe Rbeapi::Client do
 
   describe '#reload' do
     it 'reloads the configuration file' do
-      expect(subject.load_config(dut_conf))
+      expect(subject.config.get_connection('veos01')).to eq(veos01)
+      expect(subject.config.reload(filename: [dut_conf]))
         .to eq(transport: 'socket')
-      expect(subject.config.reload(filename: [test_conf]))
-        .to eq(nil)
-      expect(subject.config.to_s).to include(test_data[0])
+      expect(subject.config.get_connection('veos01')).to eq(nil)
+      expect(subject.config.get_connection('dut')).not_to be_nil
     end
   end
 
@@ -143,19 +184,38 @@ describe Rbeapi::Client do
                                                    password: 'test',
                                                    transport: 'http',
                                                    host: 'test2')
+      expect(subject.config.get_connection('test2'))
+        .to eq(username: 'test2',
+               password: 'test',
+               transport: 'http',
+               host: 'test2')
     end
   end
 
   # Node Class Methods
+  describe 'node' do
+    it 'retrieves the node' do
+      expect(node).to be_kind_of(Rbeapi::Client::Node)
+    end
+  end
+
   describe '#running_config' do
     it 'gets the nodes running config' do
-      expect(node.running_config).not_to be_nil
+      expect(node.running_config).to be_truthy
+    end
+
+    it 'expects running config to return a string' do
+      expect(node.running_config).to be_kind_of(String)
     end
   end
 
   describe '#startup_config' do
     it 'gets the nodes startup-configuration' do
-      expect(node.startup_config).not_to be_nil
+      expect(node.startup_config).to be_truthy
+    end
+
+    it 'expects startup-configuration to be a string' do
+      expect(node.startup_config).to be_kind_of(String)
     end
   end
 
@@ -167,46 +227,81 @@ describe Rbeapi::Client do
 
   describe '#config' do
     it 'puts switch into config mode' do
-      expect(node.config(['no ip virtual-router mac-address'])).to be_truthy
+      expect(node.config(['no ip virtual-router mac-address']))
+        .to be_truthy
     end
 
-    it 'puts switch into config mode with options' do
+    it 'expects config to return array' do
+      expect(node.config(['no ip virtual-router mac-address']))
+        .to be_kind_of(Array)
+    end
+
+    it 'puts switch into config mode with options and returns array' do
       expect(node.config(['no ip virtual-router mac-address'],
                          encoding: 'json',
                          open_timeout: 27.00,
-                         read_timeout: 27.00)).to be_truthy
+                         read_timeout: 27.00))
+        .to be_kind_of(Array)
+    end
+
+    describe 'set dry run' do
+      before do
+        # Prevents puts from writing to console
+        allow($stdout).to receive(:puts)
+        node.dry_run = true
+      end
+
+      it 'expects config to do dry run' do
+        expect(node.config(['no ip virtual-router mac-address']))
+          .to eq(nil)
+      end
+    end
+
+    it 'returns error if invalid command' do
+      expect { node.config(['no ip virtual-router mac-addresses']) }
+        .to raise_error Rbeapi::Eapilib::CommandError
     end
   end
 
   describe '#enable' do
     it 'puts the switch into privilege mode' do
-      expect(node.enable('show hostname')[0])
-        .to include(:command, :result, :encoding)
+      expect(node.enable('show hostname')[0][:result])
+        .to include('fqdn', 'hostname')
     end
 
     it 'puts the switch into privilege mode with encoding' do
-      expect(node.enable('show hostname', encoding: 'text')[0])
+      expect(node.enable('show hostname', encoding: 'text')[0][:encoding])
+        .to eq('text')
+    end
+
+    it 'puts the switch into privilege mode with strict option' do
+      expect(node.enable('show hostname', strict: true)[0])
         .to include(:command, :result, :encoding)
     end
 
-    it 'puts the switch into privilege mode with open and read timeout' do
+    it 'puts the switch into privilege mode with read and open timeout' do
       expect(node.enable('show hostname',
                          open_timeout: 29,
                          read_timeout: 29)[0]).to include(:command,
                                                           :result,
                                                           :encoding)
     end
+
+    it 'raises invalid command error' do
+      expect { node.enable(['show hostname', 'do this thing']) }
+        .to raise_error Rbeapi::Eapilib::CommandError
+    end
   end
 
   describe '#run_commands' do
-    it 'sends commands to node' do
-      expect(node.run_commands(['show hostname'])[0])
-        .to include('fqdn', 'hostname')
+    it 'expects run_commands to be a string' do
+      expect(node.run_commands('show hostname', encoding: 'text')[0]['output'])
+        .to be_kind_of String
     end
 
     it 'sends commands to node with encoding' do
       expect(node.run_commands('show hostname', encoding: 'text')[0]['output'])
-        .to include('FQDN', 'Hostname')
+        .to include('FQDN:', 'Hostname:')
     end
 
     it 'sends commands with open and read timeout' do
@@ -215,11 +310,42 @@ describe Rbeapi::Client do
                                read_timeout: 26)[0]).to include('fqdn',
                                                                 'hostname')
     end
+
+    it 'sends commands with enablepwd set' do
+      expect(node.enable_authentication('icanttellyou')).to eq('icanttellyou')
+      expect(node.run_commands('show hostname')).to be_truthy
+    end
+
+    it 'expects run_commands to raise a command error' do
+      expect { node.run_commands('do this thing') }
+        .to raise_error Rbeapi::Eapilib::CommandError
+    end
+  end
+
+  describe '#get_config' do
+    it 'will retrieve the specified configuration and return array' do
+      expect(node.get_config(config: 'running-config'))
+        .to be_kind_of(Array)
+    end
+
+    it 'will retrieve with param and return array' do
+      expect(node.get_config(config: 'running-config', param: 'all'))
+        .to be_kind_of(Array)
+    end
+
+    it 'raises invalid command error' do
+      expect { node.get_config(config: 'running-configurations') }
+        .to raise_error Rbeapi::Eapilib::CommandError
+    end
   end
 
   describe '#api' do
     it 'returns api module' do
       expect(node.api('vlans')).to be_kind_of(Rbeapi::Api::Vlans)
+    end
+
+    it 'returns error if invalid name' do
+      expect { node.api('vlanss') }.to raise_error
     end
   end
